requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gevent==21.8.0
greenlet==1.1.1
idna==3.2
kombu==5.1.0
lxml==4.6.3
prompt-toolkit==3.0.19
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyodbc==4.0.31
pytz==2021.1
redis==3.5.3
requests==2.26.0
rest-condition==1.0.3
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.4.0
